switch Token from uint16 to uint32

+++ pkg/lexing/token/token.go
@@ -18,13 +18,13 @@ func (t Token) String() string {
 }
 
 func (t *Token) SetStart(inputPosition int, textPosition position.Position) {
-	t.Literal.Start = uint16(inputPosition)
+	t.Literal.Start = uint32(inputPosition)
 	t.TextPosition.LineStart = textPosition.LineStart
 	t.TextPosition.CharStart = textPosition.CharStart
 }
 
 func (t *Token) SetEnd(inputPosition int, textPosition position.Position) {
-	t.Literal.End = uint16(inputPosition)
+	t.Literal.End = uint32(inputPosition)
 	t.TextPosition.LineEnd = textPosition.LineStart
 	t.TextPosition.CharEnd = textPosition.CharStart
 }
